refactor: 예약 정보 domain, entity, dto 수정
- manager 전체 정보 대신, manager id 전달
