+++ app/src/main/java/com/example/team25/data/network/dto/ReservationDto.kt
@@ -1,17 +1,17 @@
 package com.example.team25.data.network.dto
 
 import com.example.team25.domain.model.Patient
-import com.example.team25.domain.model.ReservationStatus
+import com.example.team25.domain.ReservationStatus
 import com.google.gson.annotations.SerializedName
 
 data class ReservationDto(
-    @SerializedName("manager") val manager: ManagerDto,
+    @SerializedName("managerId") val managerId: String,
     @SerializedName("reservationStatus") val reservationStatus: ReservationStatus,
-    @SerializedName("departure")val departure: String,
-    @SerializedName("destination")val destination: String,
-    @SerializedName("serviceDate")val serviceDate: String,
-    @SerializedName("serviceType")val serviceType: String,
-    @SerializedName("transportation")val transportation: String,
-    @SerializedName("price")val price: Int,
-    @SerializedName("patient")val patient: Patient,
+    @SerializedName("departure") val departure: String,
+    @SerializedName("destination") val destination: String,
+    @SerializedName("serviceDate") val serviceDate: String,
+    @SerializedName("serviceType") val serviceType: String,
+    @SerializedName("transportation") val transportation: String,
+    @SerializedName("price") val price: Int,
+    @SerializedName("patient") val patient: Patient,
 )
